fix: seperate each service for naver
