--- src/controllers/connector/naver/NaverModule.ts
@@ -1,8 +1,0 @@
-import { Module } from "@nestjs/common";
-import { NaverController } from "./NaverController";
-
-@Module({
-  controllers: [NaverController],
-  providers: [],
-})
-export class NaverModule {}
