+++ src/controllers/connector/naver_blog/NaverBlogModule.ts
@@ -1,0 +1,8 @@
+import { Module } from "@nestjs/common";
+import { NaverBlogController } from "./NaverBlogController";
+
+@Module({
+  controllers: [NaverBlogController],
+  providers: [],
+})
+export class NaverBlogModule {}
