+++ src/controllers/connector/naver_cafe/NaverCafeModule.ts
@@ -1,0 +1,8 @@
+import { Module } from "@nestjs/common";
+import { NaverCafeController } from "./NaverCafeController";
+
+@Module({
+  controllers: [NaverCafeController],
+  providers: [],
+})
+export class NaverCafeModule {}
