+++ src/controllers/connector/naver_news/NaverNewsModule.ts
@@ -1,0 +1,8 @@
+import { Module } from "@nestjs/common";
+import { NaverNewsController } from "./NaverNewsController";
+
+@Module({
+  controllers: [NaverNewsController],
+  providers: [],
+})
+export class NaverNewsModule {}
