+++ test/features/api/connector/naver_blog/test_api_connector_read_blog_naver.ts
@@ -1,18 +1,18 @@
 import CApi from "@wrtn/connector-api/lib/index";
-import { INaver } from "@wrtn/connector-api/lib/structures/connector/naver/INaver";
+import { INaverBlog } from "@wrtn/connector-api/lib/structures/connector/naver_blog/INaverBlog";
 import typia from "typia";
 
 export const test_api_connector_read_blog_naver = async (
   connection: CApi.IConnection,
-): Promise<INaver.IBlogNaverOutput> => {
-  const input: INaver.INaverKeywordInput = {
+): Promise<INaverBlog.IBlogNaverOutput> => {
+  const input: INaverBlog.INaverKeywordInput = {
     andKeywords: "뤼튼",
     orKeywords: "AI",
     notKeywords: "openAI",
     display: 10,
     sort: "sim",
   };
-  const result = await CApi.functional.connector.naver.blog.blogList(
+  const result = await CApi.functional.connector.naver_blog.blogList(
     connection,
     input,
   );
