+++ test/features/api/connector/naver_news/test_api_connector_read_news_naver.ts
@@ -1,14 +1,14 @@
 import CApi from "@wrtn/connector-api/lib/index";
-import { INaver } from "@wrtn/connector-api/lib/structures/connector/naver/INaver";
+import { INaverNews } from "@wrtn/connector-api/lib/structures/connector/naver_news/INaverNews";
 import typia from "typia";
 
 export const test_api_connector_read_news_naver = async (
   connection: CApi.IConnection,
-): Promise<INaver.INewsNaverOutput> => {
-  const input: INaver.INaverKeywordInput = {
+): Promise<INaverNews.INewsNaverOutput> => {
+  const input: INaverNews.INaverKeywordInput = {
     andKeywords: "뤼튼",
   };
-  const result = await CApi.functional.connector.naver.news.newsList(
+  const result = await CApi.functional.connector.naver_news.newsList(
     connection,
     input,
   );
